requirements.txt
django
djangorestframework
django-environ
pygments
drf-spectacular
django-filter

#Postgres Requirement
# sudo apt install python3-dev libpq-dev
#development
psycopg2-binary
#production ready
# psycopg2